- Moved old test files to the new location
